Moved LawnMower into Map file
Previously was in Plant alongside the other projectiles. Also renamed the old Lawnmower and ILawnmower to reflect that they are unused.
